non-multistage/requirements.txt
flask==2.3.2
werkzeug==2.3.3
python-dotenv==0.20.0
gunicorn==20.1.0
requests==2.27.1
